requirements.txt
beautifulsoup4==4.12.0
selenium==4.10.0
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
sentence-transformers==2.2.2
pinecone-client==2.2.1
transformers==4.30.0
streamlit==1.25.0
crewai==0.1.0
python-dotenv==1.0.0